Refactor profile to sample
