+++ ddtrace/internal/datadog/profiling/dd_wrapper/CMakeLists.txt
@@ -14,9 +14,11 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 get_filename_component(dd_wrapper_BUILD_PARENT "${CMAKE_BINARY_DIR}/../.." ABSOLUTE)
 set(dd_wrapper_BUILD_DIR "${dd_wrapper_BUILD_PARENT}/ddtrace.internal.datadog.profiling")
 
+# Custom modules are in the parent directory
+list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/../cmake")
+
 # Setup state for FindLibdatadog
 set(Datadog_BUILD_DIR "${dd_wrapper_BUILD_PARENT}/libdatadog")
-list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/../cmake")
 
 # Includes
 include(CMakePackageConfigHelpers) # For generating the config file
@@ -37,26 +39,28 @@ add_compile_options(
   -Wall -Werror -Wextra -Wshadow -Wnon-virtual-dtor -Wold-style-cast
 )
 
-# If we were given flags to do sanitization or static analyis, use them
-if (SANITIZE_OPTIONS)
-  if (SANITIZE_OPTIONS MATCHES "fanalyzer")
-    add_compile_options(-fanalyzer)
-  else()
-    add_compile_options(-fsanitize=${SANITIZE_OPTIONS})
-    add_link_options(-fsanitize=${SANITIZE_OPTIONS})
-  endif()
-endif()
-
 # Library sources
 add_library(dd_wrapper SHARED
     src/uploader_builder.cpp
-    src/profile_builder.cpp
+    src/sample_builder.cpp
     src/global_cache.cpp
-    src/profile_shared.cpp
-    src/uploader.cpp
     src/profile.cpp
+    src/uploader.cpp
+    src/sample.cpp
     src/interface.cpp
 )
+
+# If we were given flags to do sanitization or static analyis, use them
+if (SANITIZE_OPTIONS)
+  if (SANITIZE_OPTIONS MATCHES "fanalyzer")
+    target_compile_options(dd_wrapper PRIVATE -fanalyzer)
+  elseif (SANITIZE_OPTIONS MATCHES "cppcheck")
+    set(DO_CPPCHECK ON)
+  else()
+    target_compile_options(dd_wrapper PRIVATE -fsanitize=${SANITIZE_OPTIONS})
+    target_link_options(dd_wrapper PRIVATE -fsanitize=${SANITIZE_OPTIONS})
+  endif()
+endif()
 
 target_include_directories(dd_wrapper PRIVATE
     include
@@ -81,6 +85,33 @@ message(STATUS "dd_wrapper_INSTALL_DIR: ${dd_wrapper_INSTALL_DIR}")
 install(TARGETS dd_wrapper
     DESTINATION ${dd_wrapper_INSTALL_DIR}
 )
+
+# if "DO_CPPCHECK" is set, then we'll run cppcheck
+if (DO_CPPCHECK)
+  message(STATUS "Running cppcheck")
+  include(FindCppcheck)
+  if (CPPCHECK_EXECUTABLE)
+    message(STATUS "cppcheck found: ${CPPCHECK_EXECUTABLE}")
+    add_custom_target(cppcheck_run
+      COMMAND ${CPPCHECK_EXECUTABLE}
+      --enable=warning,performance,portability,information,style,missingInclude
+      --template="cppcheck:{id}:{file}:{line}:{severity}:{message}"
+      --library=googletest
+      --std=c++17
+      --language=c++
+      --suppress=missingIncludeSystem
+      --inline-suppr
+      --error-exitcode=1
+      -I ${CMAKE_SOURCE_DIR}/include
+      -I ${Datadog_INCLUDE_DIRS}
+      ${CMAKE_SOURCE_DIR}/src
+      WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
+      COMMENT "Running cppcheck"
+    )
+  else()
+    message(WARNING "cppcheck not found")
+  endif()
+endif()
 
 ### Testing
 FetchContent_Declare(
