--- ddtrace/internal/datadog/profiling/dd_wrapper/include/profile_shared.hpp
@@ -1,98 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0. This product includes software
-// developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present
-// Datadog, Inc.
-
-#pragma once
-
-#include "types.hpp"
-
-#include <atomic>
-#include <deque>
-#include <memory>
-#include <mutex>
-#include <string>
-#include <string_view>
-#include <unordered_set>
-#include <vector>
-
-extern "C"
-{
-#include "datadog/profiling.h"
-}
-
-namespace Datadog {
-
-// Unordered containers don't get heterogeneous lookup until gcc-10, so for now use this
-// strategy to dedup + store strings.
-using StringTable = std::unordered_set<std::string_view>;
-
-// This is the shared state for all Profile objects  It should be marked as
-// as static in the Profile object.
-class ProfileSharedState
-{
-  private:
-    // Serialization for static state
-    // - string table
-    // - ddog_profile
-    std::atomic<bool> first_time{ true };
-    std::atomic<long> g_profile_seq{ 0 };
-    std::mutex profile_mtx{};
-    std::atomic<pid_t> current_pid{ 0 };
-
-    // Storage for strings
-    std::deque<std::string> string_storage{};
-    StringTable strings{};
-    std::mutex string_table_mtx{};
-
-    // Configuration
-    ProfileType type_mask{ 0 };
-    unsigned int max_nframes{ 128 };
-    ddog_prof_Period default_period;
-
-    // Lookup for values
-    ValueIndex val_idx{};
-
-    // Configuration for the pprof exporter
-    std::vector<ddog_prof_ValueType> samplers{};
-
-    // Intermediate storage; this gets flushed to the exporter
-    // when it's time to upload.
-    ddog_prof_Profile cur_profile;
-    ddog_prof_Profile last_profile;
-
-    // Global state
-    inline static std::atomic<bool> dirty;
-
-  public:
-    void one_time_init(ProfileType type, unsigned int _max_nframes);
-    void setup_samplers();
-    bool cycle_buffers();
-    void entrypoint_check();
-
-    // Getters
-    unsigned int get_max_nframes();
-    ProfileType get_type_mask();
-    size_t get_sample_type_length();
-    long get_profile_seq();
-    ddog_prof_Profile& get_current_profile();
-
-    // String table manipulation
-    std::string_view insert_or_get(std::string_view sv);
-
-    // Value index getter
-    const ValueIndex& val();
-
-    // Flush
-    bool flush_sample(ddog_prof_Sample& sample);
-
-    // Global manipulators
-    static void mark_dirty();
-};
-
-struct DdogProfExporterDeleter
-{
-    void operator()(ddog_prof_Exporter* ptr) const;
-};
-
-} // namespace Datadog
