+++ ddtrace/internal/datadog/profiling/dd_wrapper/include/sample.hpp
@@ -1,0 +1,97 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0. This product includes software
+// developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present
+// Datadog, Inc.
+
+#pragma once
+
+#include "libdatadog_helpers.hpp"
+#include "profile.hpp"
+#include "types.hpp"
+
+#include <array>
+#include <string>
+#include <string_view>
+#include <vector>
+
+extern "C"
+{
+#include "datadog/profiling.h"
+}
+
+namespace Datadog {
+
+class Sample
+{
+  private:
+    static inline Profile profile_state{};
+    long profile_seq; // TODO is this being used properly?
+    unsigned int max_nframes;
+    SampleType type_mask;
+    std::string errmsg;
+
+    // Keeps temporary buffer of frames in the stack
+    std::vector<ddog_prof_Location> locations;
+
+    // Storage for labels
+    std::array<ddog_prof_Label, static_cast<size_t>(ExportLabelKey::_Length)> labels{};
+    size_t cur_label = 0;
+
+    // Storage for values
+    std::vector<int64_t> values = {};
+
+    // Helpers
+    bool push_label(const ExportLabelKey key, std::string_view val);
+    bool push_label(const ExportLabelKey key, int64_t val);
+    void push_frame_impl(std::string_view name, std::string_view filename, uint64_t address, int64_t line);
+
+  public:
+    uint64_t samples = 0;
+
+    // Getters
+    static ddog_prof_Profile& get_ddog_profile();
+
+    // Clears the current sample without flushing
+    static Sample& start_sample();
+
+    // Add values
+    bool push_walltime(int64_t walltime, int64_t count);
+    bool push_cputime(int64_t cputime, int64_t count);
+    bool push_acquire(int64_t acquire_time, int64_t count);
+    bool push_release(int64_t lock_time, int64_t count);
+    bool push_alloc(uint64_t size, uint64_t count);
+    bool push_heap(uint64_t size);
+
+    // Adds metadata to sample
+    bool push_lock_name(std::string_view lock_name);
+    bool push_threadinfo(int64_t thread_id, int64_t thread_native_id, std::string_view thread_name);
+    bool push_task_id(int64_t task_id);
+    bool push_task_name(std::string_view task_name);
+    bool push_span_id(uint64_t span_id);
+    bool push_local_root_span_id(uint64_t local_root_span_id);
+    bool push_trace_type(std::string_view trace_type);
+    bool push_trace_resource_container(std::string_view trace_resource_container);
+    bool push_exceptioninfo(std::string_view exception_type, int64_t count);
+    bool push_class_name(std::string_view class_name);
+
+    // Assumes frames are pushed in leaf-order
+    void push_frame(std::string_view name,     // for ddog_prof_Function
+                    std::string_view filename, // for ddog_prof_Function
+                    uint64_t address,          // for ddog_prof_Location
+                    int64_t line               // for ddog_prof_Location
+    );
+
+    // Flushes the current buffer, clearing it
+    bool flush_sample();
+
+    // Clears temporary things
+    void clear_stringtable();
+    void clear_buffers();
+
+    void zero_stats();
+
+    Sample(SampleType type, unsigned int _max_nframes);
+    ~Sample() = default;
+};
+
+} // namespace Datadog
