+++ ddtrace/internal/datadog/profiling/dd_wrapper/include/uploader.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "profile.hpp"
+#include "sample.hpp"
 #include "types.hpp"
 
 extern "C"
@@ -14,6 +14,11 @@ extern "C"
 }
 
 namespace Datadog {
+
+struct DdogProfExporterDeleter
+{
+    void operator()(ddog_prof_Exporter* ptr) const;
+};
 
 class Uploader
 {
