+++ ddtrace/internal/datadog/profiling/dd_wrapper/src/global_cache.cpp
@@ -5,48 +5,48 @@
 
 using namespace Datadog;
 
-ProfileGlobalStorage&
-ProfileGlobalStorage::get_singleton()
+GlobalCache&
+GlobalCache::get_singleton()
 {
-    static ProfileGlobalStorage instance;
+    static GlobalCache instance;
     return instance;
 }
 
-Profile&
-ProfileGlobalStorage::get(std::thread::id id)
+Sample&
+GlobalCache::get(std::thread::id id)
 {
-    ProfileGlobalStorage& global_store = get_singleton();
+    GlobalCache& global_store = get_singleton();
 
-    std::unique_lock<std::mutex> lock(global_store.profile_storage_mtx);
-    auto& profiles = global_store.profile_cache;
-    auto it = profiles.find(id); // Attempt to find the profile once
+    std::unique_lock<std::mutex> lock(global_store.sample_storage_mtx);
+    auto& samples = global_store.sample_cache;
+    auto it = samples.find(id); // Attempt to find the sample once
 
-    if (it == profiles.end()) {
-        // Profile not found, unlock the mutex while creating the profile
+    if (it == samples.end()) {
+        // Sample not found, unlock the mutex while creating the sample
         lock.unlock();
 
-        auto new_profile = global_store.profile_builder.build();
-        if (!new_profile) {
-            throw std::runtime_error("Error creating profile");
+        auto new_sample = global_store.sample_builder.build();
+        if (!new_sample) {
+            throw std::runtime_error("Error creating sample");
         }
 
-        // Lock again to insert the new profile
+        // Lock again to insert the new sample
         lock.lock();
-        auto [new_it, inserted] = profiles.emplace(
-          std::piecewise_construct, std::forward_as_tuple(id), std::forward_as_tuple(std::move(new_profile.value())));
+        auto [new_it, inserted] = samples.emplace(
+          std::piecewise_construct, std::forward_as_tuple(id), std::forward_as_tuple(std::move(new_sample.value())));
         return new_it->second;
     }
 
-    // If we're here, it's because the profile was found, so just return it
+    // If we're here, it's because the sample was found, so just return it
     return it->second;
 }
 
 void
-ProfileGlobalStorage::clear()
+GlobalCache::clear()
 {
     // It's UB, but testing shows that on Linux, the pthreads implementation
     // backing std::mutex guarantees consistency across forks.
-    ProfileGlobalStorage& global_store = get_singleton();
-    std::lock_guard<std::mutex> lock(global_store.profile_storage_mtx);
-    global_store.profile_cache.clear();
+    GlobalCache& global_store = get_singleton();
+    std::lock_guard<std::mutex> lock(global_store.sample_storage_mtx);
+    global_store.sample_cache.clear();
 }
