+++ ddtrace/internal/datadog/profiling/dd_wrapper/src/interface.cpp
@@ -6,8 +6,8 @@
 #include "interface.hpp"
 #include "global_cache.hpp"
 #include "libdatadog_helpers.hpp"
+#include "sample.hpp"
 #include "profile.hpp"
-#include "profile_shared.hpp"
 #include "uploader.hpp"
 
 #include <cstdlib>
@@ -20,15 +20,15 @@ bool g_prof_flag = true;
 
 // Store a pointer to the current profile.  This can be assumed set in inner
 // functions, but any function that may be called by a new thread should check
-thread_local Datadog::Profile* g_profile;
+thread_local Datadog::Sample* g_profile;
 
 // When a fork is detected, we need to reinitialize this state.
 // This handler will be called in the single thread of the child process after the fork
 void
 ddup_fork_handler()
 {
-    Datadog::ProfileSharedState::mark_dirty();
-    Datadog::ProfileGlobalStorage::clear();
+    Datadog::Profile::mark_dirty();
+    Datadog::GlobalCache::clear();
 }
 
 // Configuration
@@ -38,7 +38,7 @@ ddup_config_env(const char* env)
     if (!env || !*env)
         return;
 
-    Datadog::ProfileGlobalStorage::uploader_builder.set_env(env);
+    Datadog::GlobalCache::uploader_builder.set_env(env);
 }
 void
 ddup_config_service(const char* service)
@@ -47,60 +47,60 @@ ddup_config_service(const char* service)
         return;
     }
 
-    Datadog::ProfileGlobalStorage::uploader_builder.set_service(service);
+    Datadog::GlobalCache::uploader_builder.set_service(service);
 }
 void
 ddup_config_version(const char* version)
 {
     if (!version || !*version)
         return;
-    Datadog::ProfileGlobalStorage::uploader_builder.set_version(version);
+    Datadog::GlobalCache::uploader_builder.set_version(version);
 }
 void
 ddup_config_runtime(const char* runtime)
 {
     if (!runtime || !*runtime)
         return;
-    Datadog::ProfileGlobalStorage::uploader_builder.set_runtime(runtime);
+    Datadog::GlobalCache::uploader_builder.set_runtime(runtime);
 }
 void
 ddup_config_runtime_version(const char* runtime_version)
 {
     if (!runtime_version || !*runtime_version)
         return;
-    Datadog::ProfileGlobalStorage::uploader_builder.set_runtime_version(runtime_version);
+    Datadog::GlobalCache::uploader_builder.set_runtime_version(runtime_version);
 }
 void
 ddup_config_profiler_version(const char* profiler_version)
 {
     if (!profiler_version || !*profiler_version)
         return;
-    Datadog::ProfileGlobalStorage::uploader_builder.set_profiler_version(profiler_version);
+    Datadog::GlobalCache::uploader_builder.set_profiler_version(profiler_version);
 }
 void
 ddup_config_url(const char* url)
 {
     if (!url || !*url)
         return;
-    Datadog::ProfileGlobalStorage::uploader_builder.set_url(url);
+    Datadog::GlobalCache::uploader_builder.set_url(url);
 }
 void
 ddup_config_user_tag(const char* key, const char* val)
 {
     if (!key || !*key || !val || !*val)
         return;
-    Datadog::ProfileGlobalStorage::uploader_builder.set_tag(key, val);
+    Datadog::GlobalCache::uploader_builder.set_tag(key, val);
 }
 void
 ddup_config_sample_type(unsigned int type)
 {
-    Datadog::ProfileGlobalStorage::profile_builder.add_type(type);
+    Datadog::GlobalCache::sample_builder.add_type(type);
 }
 void
 ddup_config_max_nframes(int max_nframes)
 {
     if (max_nframes > 0)
-        Datadog::ProfileGlobalStorage::profile_builder.set_max_nframes(max_nframes);
+        Datadog::GlobalCache::sample_builder.set_max_nframes(max_nframes);
 }
 
 #if DDUP_BACKTRACE_ENABLE
@@ -185,7 +185,7 @@ ddup_init()
 void
 ddup_start_sample()
 {
-    g_profile = &Datadog::Profile::start_sample();
+    g_profile = &Datadog::Sample::start_sample();
 }
 
 void
@@ -312,7 +312,7 @@ void
 ddup_set_runtime_id(const char* id, size_t sz)
 {
     if (id && *id)
-        Datadog::ProfileGlobalStorage::uploader_builder.set_runtime_id(std::string_view(id, sz));
+        Datadog::GlobalCache::uploader_builder.set_runtime_id(std::string_view(id, sz));
 }
 
 bool
@@ -323,11 +323,11 @@ ddup_upload()
         return false;
     }
 
-    ddog_prof_Profile upload_profile = Datadog::Profile::get_ddog_profile();
-    Datadog::ProfileSharedState::mark_dirty();
+    ddog_prof_Profile upload_profile = Datadog::Sample::get_ddog_profile();
+    Datadog::Profile::mark_dirty();
 
     // We create a new uploader just for this operation
-    auto uploader = Datadog::ProfileGlobalStorage::uploader_builder.build_ptr();
+    auto uploader = Datadog::GlobalCache::uploader_builder.build_ptr();
     if (!uploader) {
         std::cerr << "Failed to create uploader" << std::endl;
         return false;
